Exporters refactoring + DXF spline/any curve support
