+++ cadquery/occ_impl/exporters/utils.py
@@ -1,0 +1,7 @@
+from ...cq import Workplane
+from ..shapes import Compound, Shape
+
+
+def toCompound(shape: Workplane) -> Compound:
+
+    return Compound.makeCompound(val for val in shape.vals() if isinstance(val, Shape))
